+++ CHANGELOG.md
@@ -5,6 +5,9 @@ FlatLaf Change Log
 
 #### New features and improvements
 
+- TextField, FormattedTextField and PasswordField: Support leading and trailing
+  icons (set client property `JTextField.leadingIcon` or
+  `JTextField.trailingIcon` to an `Icon`). (issue #368)
 - InternalFrame: Double-click on icon in internal frame title bar now closes the
   internal frame. (issue #374)
 - IntelliJ Themes: Removed deprecated `install()` methods.
@@ -55,8 +58,8 @@ FlatLaf Change Log
 
 #### New features and improvements
 
-- TextField, FormattedTextField and PasswordField: Support adding extra padding.
-  (set client property `JTextField.padding` to `Insets`).
+- TextField, FormattedTextField and PasswordField: Support adding extra padding
+  (set client property `JTextField.padding` to an `Insets`).
 - PasswordField: UI delegate `FlatPasswordFieldUI` now extends `FlatTextFieldUI`
   (instead of `BasicPasswordFieldUI`) to avoid duplicate code and for easier
   extensibility.